Day-31_18-11/airflow_dags/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
pyarrow
sqlalchemy
psycopg2-binary